Map Ops now works: attrs_sz now determined

+++ hyperclient/java/extra_src/DeferredFromAttrs.java
@@ -13,7 +13,7 @@ public class DeferredFromAttrs extends Deferred
         super(client);
 
 	      hyperclient_attribute attrs = null;
-	      int attrs_sz = 0;
+	      long attrs_sz = 0;
 	
         try
         {
