+++ hyperclient/java/extra_src/MapOpAdd.java
@@ -10,7 +10,7 @@ class MapOpAdd extends MapOp
     }
 
     long call(String space, String key,
-              hyperclient_map_attribute attrs, int attrs_sz,
+              hyperclient_map_attribute attrs, long attrs_sz,
               SWIGTYPE_p_int rc_int_ptr)
     {
         return client.map_add(space, key, attrs, attrs_sz, rc_int_ptr);
